Updated Project

+++ crud-angular/src/app/courses/containers/courses/courses.component.ts
@@ -1,22 +1,44 @@
-import { Component, OnInit } from '@angular/core';
-import { MatDialog } from '@angular/material/dialog';
-import { MatSnackBar } from '@angular/material/snack-bar';
+import { AsyncPipe, NgIf } from '@angular/common';
+import { Component, OnInit, ViewChild } from '@angular/core';
+import { MatCardModule } from '@angular/material/card';
+import { MatDialog, MatDialogModule } from '@angular/material/dialog';
+import { MatPaginator, MatPaginatorModule, PageEvent } from '@angular/material/paginator';
+import { MatProgressSpinnerModule } from '@angular/material/progress-spinner';
+import { MatSnackBar, MatSnackBarModule } from '@angular/material/snack-bar';
+import { MatToolbarModule } from '@angular/material/toolbar';
 import { ActivatedRoute, Router } from '@angular/router';
-import { Observable, of } from 'rxjs';
-import { catchError } from 'rxjs/operators';
+import { catchError, Observable, of, tap } from 'rxjs';
 
+import { ConfirmationDialogComponent } from '../../../shared/components/confirmation-dialog/confirmation-dialog.component';
 import { ErrorDialogComponent } from '../../../shared/components/error-dialog/error-dialog.component';
-import { Course } from '../../models/course';
+import { CoursesListComponent } from '../../components/courses-list/courses-list.component';
+import { Course } from '../../model/course';
+import { CoursePage } from '../../model/course-page';
 import { CoursesService } from '../../services/courses.service';
 
 @Component({
-  selector: 'app-courses',
-  templateUrl: './courses.component.html',
-  styleUrls: ['./courses.component.scss']
+    selector: 'app-courses',
+    templateUrl: './courses.component.html',
+    styleUrls: ['./courses.component.scss'],
+    imports: [
+        MatCardModule,
+        MatToolbarModule,
+        NgIf,
+        CoursesListComponent,
+        MatProgressSpinnerModule,
+        MatSnackBarModule,
+        MatDialogModule,
+        MatPaginatorModule,
+        AsyncPipe
+    ]
 })
 export class CoursesComponent implements OnInit {
+  courses$: Observable<CoursePage> | null = null;
 
-  courses$: Observable<Course[]> | null = null;
+  pageIndex = 0;
+  pageSize = 10;
+
+  @ViewChild(MatPaginator) paginator!: MatPaginator;
 
   constructor(
     private coursesService: CoursesService,
@@ -24,16 +46,23 @@ export class CoursesComponent implements OnInit {
     private router: Router,
     private route: ActivatedRoute,
     private snackBar: MatSnackBar
-  ) {
+  ) { }
+
+  ngOnInit() {
     this.refresh();
   }
 
-  refresh() {
-    this.courses$ = this.coursesService.list()
+  refresh(pageEvent: PageEvent = { length: 0, pageIndex: 0, pageSize: 10 }) {
+    this.courses$ = this.coursesService
+      .list(pageEvent.pageIndex, pageEvent.pageSize)
       .pipe(
-        catchError(error => {
-          this.onError('Erro ao carregar cursos.');
-          return of([])
+        tap(() => {
+          this.pageIndex = pageEvent.pageIndex;
+          this.pageSize = pageEvent.pageSize;
+        }),
+        catchError(() => {
+          this.onError('Error loading courses.');
+          return of({ courses: [], totalElements: 0 } as CoursePage);
         })
       );
   }
@@ -44,8 +73,6 @@ export class CoursesComponent implements OnInit {
     });
   }
 
-  ngOnInit(): void { }
-
   onAdd() {
     this.router.navigate(['new'], { relativeTo: this.route });
   }
@@ -54,18 +81,29 @@ export class CoursesComponent implements OnInit {
     this.router.navigate(['edit', course._id], { relativeTo: this.route });
   }
 
-  onRemove(course: Course) {
-    this.coursesService.remove(course._id).subscribe(
-      () => {
-        this.refresh();
-        this.snackBar.open('Curso removido com sucesso!', 'X', {
-          duration: 5000,
-          verticalPosition: 'top',
-          horizontalPosition: 'center'
-        });
-      },
-      () => this.onError('Erro ao tentar remover curso.')
-    );
+  onView(course: Course) {
+    this.router.navigate(['view', course._id], { relativeTo: this.route });
   }
 
+  onRemove(course: Course) {
+    const dialogRef = this.dialog.open(ConfirmationDialogComponent, {
+      data: 'Are you sure you would like to remove this course?'
+    });
+
+    dialogRef.afterClosed().subscribe((result: boolean) => {
+      if (result) {
+        this.coursesService.remove(course._id).subscribe({
+          next: () => {
+            this.refresh();
+            this.snackBar.open('Course removed successfully!', 'X', {
+              duration: 5000,
+              verticalPosition: 'top',
+              horizontalPosition: 'center'
+            });
+          },
+          error: () => this.onError('Error trying to remove the course.')
+        });
+      }
+    });
+  }
 }
